Better role management + data init fixes

+++ VideoArchiver/WebApp/Areas/Admin/Controllers/UserManagementController.cs
@@ -1,4 +1,3 @@
-using App.BLL.DTO;
 using App.BLL.DTO.Exceptions;
 using App.BLL.Identity;
 using App.Common;
@@ -6,11 +5,12 @@ using Base.WebHelpers;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using WebApp.Areas.Admin.ViewModels;
+using WebApp.Authorization;
 
 namespace WebApp.Areas.Admin.Controllers;
 
 [Area("Admin")]
-[Authorize(Roles = RoleNames.Admin)]
+[Authorize(Roles = RoleNames.AdminOrSuperAdmin)]
 public class UserManagementController : Controller
 {
     private readonly IdentityUow _identityUow;
@@ -27,10 +27,21 @@ public class UserManagementController : Controller
     {
         return View(new UserManagementViewModel
         {
-            Users = (await _identityUow.UserService.GetUsersWithRoles(includeOnlyRequiringApproval: IncludeOnlyNotApproved, nameQuery: NameQuery))
+            Users = (await _identityUow.UserService.GetUsersWithRoles(
+                    includeOnlyRequiringApproval: IncludeOnlyNotApproved, nameQuery: NameQuery))
                 .Where(u => u.Id != User.GetUserId()).ToList(),
             IncludeOnlyNotApproved = IncludeOnlyNotApproved,
             NameQuery = NameQuery,
+        });
+    }
+
+    public async Task<IActionResult> ManageRoles(Guid userId)
+    {
+        var user = await _identityUow.UserService.GetUserWithRoles(userId);
+        if (user == null) return NotFound();
+        return View(new ManageRolesViewModel
+        {
+            User = user
         });
     }
 
@@ -47,22 +58,28 @@ public class UserManagementController : Controller
         }
 
         await _identityUow.SaveChangesAsync();
-        return RedirectToAction(nameof(Index));
+        return RedirectToAction(nameof(Index), new
+        {
+            NameQuery,
+            IncludeOnlyNotApproved,
+        });
     }
 
     [HttpPost]
     public async Task<IActionResult> AddRole(Guid userId, string roleName)
     {
+        if (!User.IsAllowedToManageRole(roleName)) return Forbid();
         await _identityUow.UserService.AddUserToRole(userId, roleName);
         await _identityUow.SaveChangesAsync();
-        return RedirectToAction(nameof(Index));
+        return RedirectToAction(nameof(ManageRoles), new { userId });
     }
 
     [HttpPost]
     public async Task<IActionResult> RemoveRole(Guid userId, string roleName)
     {
+        if (!User.IsAllowedToManageRole(roleName)) return Forbid();
         await _identityUow.UserService.RemoveUserFromRole(userId, roleName);
         // No SaveChanges, this already calls ExecuteDeleteAsync(). Reassess if not using EF Core.
-        return RedirectToAction(nameof(Index));
+        return RedirectToAction(nameof(ManageRoles), new { userId });
     }
 }
